Fix issue with binder caching in kotlin module (#19)
The cached `kotlinBinder` in `KotlinModule` would not be invalidated when multiple calls were made to the `#configure(Binder)` method on `AbstractModule` with different `Binder` instances. 

This change keeps the caching within the `#configure` method and across multiple calls to the `#configure(Binder)` method with the same `Binder` instance, and it now invalidates the cached `kotlinBinder` for different `Binder` instances.